fix: resolve all clippy warnings across lib and tests
- Add missing events/sinks fields to LinksConfig in GraphQL tests
- Collapse nested if-let statements in gRPC event_service filters
- Remove unnecessary u64→u64 cast
- Simplify map_or to is_some_and in fan_out/resolve operators
- Factor complex type alias in test sink (DeliveryLog)
- Remove unnecessary let binding in webhook test
- Remove unnecessary usize casts in memory event log test

+++ src/events/memory.rs
@@ -557,9 +557,9 @@ mod tests {
 
         assert_eq!(
             events.len(),
-            event_count as usize,
+            event_count,
             "lost {} events",
-            event_count as usize - events.len()
+            event_count - events.len()
         );
 
         // Verify sequential order and seq_no
